GH#31196: separate TOML edits and address review findings
